New version of the code. Has updated

+++ income_delete.php
@@ -1,8 +1,11 @@
 <?php
 include 'db.php';
-print_r($_GET);
-$id = $_GET['id'];
-$user_id = $_GET['user_id'];
+if (!isset($_GET['id'])){
+  header("Location: index.php");
+  exit();
+}
+$id = $_GET['i_id'];
+$user_id = $_GET['id'];
 $sql = "DELETE FROM income where i_id=$id";
 if($conn->query($sql) === TRUE){
         header("Location: income.php?id=$user_id");
